test: add test verifying the unit test outputs folders only based on the log
the alternative would be it too aggressively reporting rule workspaces based on the parsed snakemake
file. this doesn't cover the situation where the tests are installed in a directory in which a
different ruleset was previously emitted. future functionality of the primary software may cause it
to clean up after itself in that case.

re #18