test: add test for basic snakemake file inclusion
this was already theoretically working internally, as that's how the original file is included, but
it's good to confirm it works with external file specification. this specifically tests include
directives in the top-level snakefile in global scope operating directly on strings, not variables.
all those other variants need separate testing

re #18