fix: capture python exit code for subprocess error checking
this prevents infinite loops when python finds syntax it doesn't like. this will almost certainly be
caused by snakemake-specific features that are not being correctly interpreted or handled by
snakemake_unit_tests, or python structures that the hackjob interpreter is not handling correctly.
the error messages for python run failures elaborate on this slightly, and specifically request that
log data be posted to the snakemake_unit_tests repo for feedback and patching.

re #21