fix: support and test support for rule inheritance
a very old and rather deprecated solution for handling rule inheritance was previously implemented.
it was also apparently broken and completely untested. thus now it is removed and replaced with a
superior solution that is consistent with the more recent implementation choices involving inclusion
of multiple layers of rules in some cases. there are also tests, and multiple layers of parents are
now finally supported.