feat: implement better python parser for comments and strings and literals
the existing code for handling this was completely inadequate. now there is a parser, which isn't
plugged in yet, which preprocesses the code of a file and correctly merges together consecutive
logically grouped lines, unambiguously removes comments, and just generally is much better. this
should pretty much gut the existing remove_comments_and_docstrings

re #15