feat: make test output runs persist after failed runs, and provide a helper script to run them
this should hopefully make it easier to debug tests that suddenly fail. pytest_runner.bash makes the
pytest output slightly uglier but does nice maintenance. it runs either with no arguments, in which
case it runs tests for each rule in turn, or it accepts names of rules, and it runs tests for just
those rules