fix: handle python configurations that bury blockers in initially unevaluated chunks
the logic for detecting the need for iterative python updates was not sufficiently complex to be
able to detect certain conditional statements. now, each block maintains a status indicator of
whether it has ever been emitted to the python tracking system. this, combined with the running
resolution status indicator, allows an accurate assessment of blocker status even within unevaluated
code blocks on a previous passn