feat: respect ordering of rule block entries as encountered in input
in a very old version of the program, there was an intention to directly manipulate the contents of
individual parts of rules. that was long ago deemed infeasible and somewhat in contradiction to the
spirit of the objective of the program. back in that period, the reporting of block entries didn't
necessarily respect the order of the entries in the input file, which introduced problems when
certain block entries have enforced orderings in snakemake. that has been fixed now, and block
entries are reported out in exactly the same order in which they're encountered. this also renders
the program substantially more resilient against future new features added upstream.

re #40