test: add unit tests for rule_block python logging
this function tests for expected self-reported tagging to python stream. this patch includes a minor
conceptual fix to the rule block logic itself: one include directive was trying to respect a global
indentation variable that has changed meaning multiple times over development. in reality, the
statement was doing nothing, as the type of rule_block object that contains include directives
always has that indentation level set to 0. however, it is still clearer and more consistent to
clean up the call.

re #40