test: add unit tests for rule_block rules dot ad hoc detection
the logic used here for detecting rules. statements is completely flawed and flagged for
deprecation. in the meantime, I've updated the string matching logic to not just be complete
garbage; however, I'd expect all of this to be removed as soon as the initial pass of unit test
writing is complete

re #40