fix: set permissions of targets that already exist and remove them before copy
snakemake offers the protected() flag that causes an output file to have write permissions removed
after creation. this permission is persistent after file copy during unit test installation, and
causes problems when the program tries to update tests to reflect expanded rulesets for rules.
and/or checkpoints. before copy, the files are now added user/write and removed, to prevent this
sort of collision in all cases.

fix #64